requirements.txt
selenium>=4.3.0
Pillow==8.3.1
